fix(clerk-js): Move ERROR_CODES constant to core
To enable code splitting of the /ui dir, /core should never depend on it. It should always be the other way around.
