+++ packages/clerk-js/src/ui/components/SignIn/SignInStart.tsx
@@ -1,9 +1,10 @@
 import type { ClerkAPIError, SignInCreateParams } from '@clerk/types';
 import React from 'react';
 
+import { ERROR_CODES } from '../../../core/constants';
 import { clerkInvalidFAPIResponse } from '../../../core/errors';
 import { getClerkQueryParam } from '../../../utils';
-import { ERROR_CODES, getIdentifierControlDisplayValues, withRedirectToHomeSingleSessionGuard } from '../../common';
+import { getIdentifierControlDisplayValues, withRedirectToHomeSingleSessionGuard } from '../../common';
 import { useCoreClerk, useCoreSignIn, useEnvironment, useSignInContext } from '../../contexts';
 import { Col, descriptors, Flow, localizationKeys } from '../../customizables';
 import {
